Fix uintptr test on 32-bit architectures
The new uintptr test is failing on 32-bit architectures (like armhf)
because of:

  src/github.com/uber-go/atomic/uintptr_test.go:72:29: constant 18446744073709551615 overflows uintptr

This is known problem and the workaround is to cast the math.MaxUint64
constant to "uint64".

Closes #99

+++ uintptr_test.go
@@ -69,7 +69,7 @@ func TestUintptr(t *testing.T) {
 	t.Run("String", func(t *testing.T) {
 		// Use an integer with the signed bit set. If we're converting
 		// incorrectly, we'll get a negative value here.
-		atom := NewUintptr(uintptr(math.MaxUint64))
+		atom := NewUintptr(uintptr(uint64(math.MaxUint64)))
 		assert.Equal(t, "18446744073709551615", atom.String(),
 			"String() returned an unexpected value.")
 	})
